refactor : 바뀐 쿼리 service에 적용 #165
