+++ src/main/java/umc/th/juinjang/service/LimjangService/LimjangQueryService.java
@@ -1,15 +1,17 @@
 package umc.th.juinjang.service.LimjangService;
 
 import java.util.List;
-import umc.th.juinjang.model.dto.limjang.LimjangDetailResponseDTO;
-import umc.th.juinjang.model.dto.limjang.LimjangMainViewListResponsetDTO;
+import umc.th.juinjang.model.dto.limjang.response.LimjangDetailResponseDTO;
+import umc.th.juinjang.model.dto.limjang.response.LimjangMainViewListResponsetDTO;
 
-import umc.th.juinjang.model.dto.limjang.LimjangTotalListResponseDTO;
+import umc.th.juinjang.model.dto.limjang.response.LimjangTotalListResponseDTO;
+import umc.th.juinjang.model.dto.limjang.enums.LimjangSortOptions;
+import umc.th.juinjang.model.dto.limjang.response.LimjangsGetResponse;
 import umc.th.juinjang.model.entity.Member;
 
 public interface LimjangQueryService {
 
-  LimjangTotalListResponseDTO.TotalListDto getLimjangTotalList(Member member, String sort);
+  LimjangsGetResponse getLimjangTotalList(Member member, LimjangSortOptions sort);
 
   List<LimjangMainViewListResponsetDTO.ListDto> getLimjangMainList(Member member);
